rename comparatorlib -> libcomparator 	new file:   CMakeLists.txt 	new file:   lib/CMakeLists.txt 	new file:   lib/include/libcomparator.hpp 	new file:   lib/src/libcomparator.cpp 	new file:   test/CMakeLists.txt 	new file:   test/comparatortest.cpp 	new file:   test/test_main.cpp
