+++ libcomparator/lib/include/libcomparator.hpp
